requirements.txt
yfinance>=0.2.55
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.31.0
newsapi-python>=0.2.7 